move connector status logic to FivetranConnector
